Add page: Astro Slide boot modes
I thought this would be useful to document, especially given the differences from the Gemini and Cosmo.

It could be worth combining this with the Gemian wiki's Bootloader page (I think here could be a better home); but I've just added a link for now. It'd be good to link back from there too.