Update Astro rooting page for firmware V07, adding A/B slot instructions
I've added the new boot images, and annotated MD5 hashes for all of them. I'm now appending the Magisk version to the filename, as there was an update since the first image was produced.

Adding instructions for determining which A/B slot's in use is now necessary, since it's no longer always going to be A.