Astro rooting: Fix URLs for boot images by updating hardcoded commit SHA to the latest
Most of the links were broken. Filenames in the project & link URLs were renamed, but were not available with those names at the original commit.

I've now tested that all these image file links actually download something, and that the checksums are correct.

Avoiding this problem was one of the reasons I'd originally not hardcoded a commit SHA =P